Clean up type truncation warnings
#jira UE-116218
#rb trivial
#rnx
#preflight 61e6efb4837b79f7ccebe551

#ROBOMERGE-AUTHOR: ben.hoffman
#ROBOMERGE-SOURCE: CL 18641648 in //UE5/Release-5.0/... via CL 18641663 via CL 18641684
#ROBOMERGE-BOT: UE5 (Release-Engine-Test -> Main) (v900-18638592)

[CL 18641708 by ben hoffman in ue5-main branch]

+++ Engine/Source/Runtime/Engine/Classes/Engine/Texture2D.h
@@ -119,9 +119,9 @@ public:
 	//~ End UObject Interface.
 
 	//~ Begin UTexture Interface.
-	virtual float GetSurfaceWidth() const override { return GetSizeX(); }
-	virtual float GetSurfaceHeight() const override { return GetSizeY(); }
-	virtual float GetSurfaceDepth() const override { return 0; }
+	virtual float GetSurfaceWidth() const override { return static_cast<float>(GetSizeX()); }
+	virtual float GetSurfaceHeight() const override { return static_cast<float>(GetSizeY()); }
+	virtual float GetSurfaceDepth() const override { return 0.0f; }
 	virtual uint32 GetSurfaceArraySize() const override { return 0; }
 	virtual TextureAddress GetTextureAddressX() const override { return AddressX; }
 	virtual TextureAddress GetTextureAddressY() const override { return AddressY; }
