+++ Engine/Source/Runtime/Engine/Classes/Engine/Texture2DArray.h
@@ -53,9 +53,9 @@ public:
 	virtual void GetAssetRegistryTags(TArray<FAssetRegistryTag>& OutTags) const override;
 	virtual FString GetDesc() override;
 	virtual void GetResourceSizeEx(FResourceSizeEx& CumulativeResourceSize) override;
-	virtual float GetSurfaceWidth() const override { return GetSizeX(); }
-	virtual float GetSurfaceHeight() const override { return GetSizeY(); }
-	virtual float GetSurfaceDepth() const override { return 0; }
+	virtual float GetSurfaceWidth() const override { return static_cast<float>(GetSizeX()); }
+	virtual float GetSurfaceHeight() const override { return static_cast<float>(GetSizeY()); }
+	virtual float GetSurfaceDepth() const override { return 0.0f; }
 	virtual uint32 GetSurfaceArraySize() const override { return GetArraySize(); }
 	virtual TextureAddress GetTextureAddressX() const override { return AddressX; }
 	virtual TextureAddress GetTextureAddressY() const override { return AddressY; }
